Additional Match Data from TBA

+++ eventModify.php
@@ -203,14 +203,31 @@
 				break;
 			}
 
+			// Get Match Id
+			$tsql = "select m.id
+			           from Match m
+					  where m.gameEventId = " . $gameEventId .
+				    "   and m.type = '" . strtoupper($value["comp_level"]) . "' " .
+					"   and m.number = '" . $matchNumber . "';";
+			$results = sqlsrv_query($conn, $tsql);
+			if(!$results) 
+			{
+				echo "Search for Match " . $value["comp_level"] . $matchNumber . " failed!<br />";
+				if( ($errors = sqlsrv_errors() ) != null) {
+					foreach( $errors as $error ) {
+						echo "SQLSTATE: ".$error[ 'SQLSTATE']."<br />";
+						echo "code: ".$error[ 'code']."<br />";
+						echo "message: ".$error[ 'message']."<br />";
+					}
+				}
+				break;
+			}
+			$row = sqlsrv_fetch_array($results, SQLSRV_FETCH_ASSOC);
+			$matchId = $row['id'];
+			
 			// Delete from Team/Match, if team not part of the match
 			$tsql = "delete from TeamMatch " .
-			        " where matchId in " .
-					"      (select m.id " .
-					"         from Match m " .
-					"        where m.gameEventId = " . $gameEventId .
-				    "          and m.type = '" . strtoupper($value["comp_level"]) . "' " .
-					"          and m.number = '" . $matchNumber . "') " .
+			        " where matchId = " . $matchId .
 					"   and teamId not in " .
 					"      (select t.id " .
 					"         from Team t " .
@@ -237,76 +254,52 @@
 			
 			// Create Match/Team Cross-Reference
 			$tsql = "insert into TeamMatch (matchId, teamId, alliance, alliancePosition) " . 
-					"select m.id, t.id, 'R', 1 " .
-					"  from Team t, " .
-					"       Match m " .
+					"select " . $matchId . ", t.id, 'R', 1 " .
+					"  from Team t " .
 				    " where t.teamNumber = " . substr($value["alliances"]["red"]["team_keys"][0], 3) .
-					"   and m.gameEventId = " . $gameEventId .
-				    "   and m.type = '" . strtoupper($value["comp_level"]) . "' " .
-					"   and m.number = '" . $matchNumber . "' " .
 					"   and not exists (select 1 " .
 					"                     from TeamMatch tm " .
-					"                    where tm.matchId = m.id " .
+					"                    where tm.matchId = " . $matchId .
 					"                      and tm.teamId = t.id) " .
 					"union " .
-					"select m.id, t.id, 'R', 2 " .
-					"  from Team t, " .
-					"       Match m " .
+					"select " . $matchId . ", t.id, 'R', 2 " .
+					"  from Team t " .
 				    " where t.teamNumber = " . substr($value["alliances"]["red"]["team_keys"][1], 3) .
-					"   and m.gameEventId = " . $gameEventId .
-				    "   and m.type = '" . strtoupper($value["comp_level"]) . "' " .
-					"   and m.number = '" . $matchNumber . "' " .
 					"   and not exists (select 1 " .
 					"                     from TeamMatch tm " .
-					"                    where tm.matchId = m.id " .
+					"                    where tm.matchId = " . $matchId .
 					"                      and tm.teamId = t.id) " .
 					"union " .
-					"select m.id, t.id, 'R', 3 " .
-					"  from Team t, " .
-					"       Match m " .
+					"select " . $matchId . ", t.id, 'R', 3 " .
+					"  from Team t " .
 				    " where t.teamNumber = " . substr($value["alliances"]["red"]["team_keys"][2], 3) .
-					"   and m.gameEventId = " . $gameEventId .
-				    "   and m.type = '" . strtoupper($value["comp_level"]) . "' " .
-					"   and m.number = '" . $matchNumber . "' " .
 					"   and not exists (select 1 " .
 					"                     from TeamMatch tm " .
-					"                    where tm.matchId = m.id " .
+					"                    where tm.matchId = " . $matchId .
 					"                      and tm.teamId = t.id) " .
 					"union " .
-					"select m.id, t.id, 'B', 1 " .
-					"  from Team t, " .
-					"       Match m " .
+					"select " . $matchId . ", t.id, 'B', 1 " .
+					"  from Team t " .
 				    " where t.teamNumber = " . substr($value["alliances"]["blue"]["team_keys"][0], 3) .
-					"   and m.gameEventId = " . $gameEventId .
-				    "   and m.type = '" . strtoupper($value["comp_level"]) . "' " .
-					"   and m.number = '" . $matchNumber . "' " .
 					"   and not exists (select 1 " .
 					"                     from TeamMatch tm " .
-					"                    where tm.matchId = m.id " .
+					"                    where tm.matchId = " . $matchId .
 					"                      and tm.teamId = t.id) " .
 					"union " .
-					"select m.id, t.id, 'B', 2 " .
-					"  from Team t, " .
-					"       Match m " .
+					"select " . $matchId . ", t.id, 'B', 2 " .
+					"  from Team t " .
 				    " where t.teamNumber = " . substr($value["alliances"]["blue"]["team_keys"][1], 3) .
-					"   and m.gameEventId = " . $gameEventId .
-				    "   and m.type = '" . strtoupper($value["comp_level"]) . "' " .
-					"   and m.number = '" . $matchNumber . "' " .
 					"   and not exists (select 1 " .
 					"                     from TeamMatch tm " .
-					"                    where tm.matchId = m.id " .
+					"                    where tm.matchId = " . $matchId .
 					"                      and tm.teamId = t.id) " .
 					"union " .
-					"select m.id, t.id, 'B', 3 " .
-					"  from Team t, " .
-					"       Match m " .
+					"select " . $matchId . ", t.id, 'B', 3 " .
+					"  from Team t " .
 				    " where t.teamNumber = " . substr($value["alliances"]["blue"]["team_keys"][2], 3) .
-					"   and m.gameEventId = " . $gameEventId .
-				    "   and m.type = '" . strtoupper($value["comp_level"]) . "' " .
-					"   and m.number = '" . $matchNumber . "' " .
 					"   and not exists (select 1 " .
 					"                     from TeamMatch tm " .
-					"                    where tm.matchId = m.id " .
+					"                    where tm.matchId = " . $matchId .
 					"                      and tm.teamId = t.id);";
 			$results = sqlsrv_query($conn, $tsql);
 			if(!$results)
